requirements.txt
flask
openai
gunicorn
python-dotenv
requests
